requirements.txt
fastapi
masonite-orm
psycopg2-binary
pytest
strawberry-graphql[fastapi]
uvicorn
requests
pyudev